+++ frontend/src/app/page.tsx
@@ -1,6 +1,18 @@
 import { ChatInput } from "@/components/chat-input";
 import { fetcher } from "@/lib/fetcher";
 import { auth } from "@/lib/auth";
+import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
+
+async function isVlabAndProjectPopulated() {
+  const cookieStore = await cookies();
+  const vlabId = cookieStore.get("virtualLabID");
+  const projectId = cookieStore.get("projectID");
+
+  if (!(vlabId && projectId)) {
+    redirect("/settings");
+  }
+}
 
 async function getToolList() {
   const session = await auth();
@@ -16,6 +28,7 @@ async function getToolList() {
 }
 
 export default async function Home() {
+  await isVlabAndProjectPopulated();
   const availableTools = await getToolList();
   return (
     <div className="flex flex-col justify-center h-full">